Title: 최솟값 3개, Time: 194ms, Memory: 1MB, Status: Passed - CodeTree

+++ 241016/최솟값 3개/top-3-smallest-number.cpp
@@ -16,7 +16,7 @@ int main() {
         if (pq.size() >= 3)
         {
             vector<int> temp(3, 0);
-            int ans = 1;
+            long long ans = 1;
             for (int i = 0; i < 3; i++)
             {
                 temp[i] = pq.top();
